+++ WaveGen/Program.cs
@@ -3,7 +3,7 @@ using System.IO;
 
 namespace WaveGen
 {
-	internal delegate float GeneratorFunc(int sampleIndex, int sampleCount);
+	internal delegate double GeneratorFunc(int sampleIndex, int sampleCount);
 
 	class Program
 	{
@@ -14,10 +14,10 @@ namespace WaveGen
 			GenerateSampleFile(GenerateVocalWave, "vocal.raw", SampleCount);
 		}
 
-		static float GenerateVocalWave(int sampleIndex, int sampleCount)
+		static double GenerateVocalWave(int sampleIndex, int sampleCount)
 		{
-			float x = (float)(sampleIndex / (sampleCount - 1f)) * (float)Math.PI * 2f;
-			return (float)Math.Sin(x + Math.Sin(x + Math.Sin(x) / 3f) / 3f);
+			double x = ((double)sampleIndex / (sampleCount - 1f)) * Math.PI * 2f;
+			return Math.Sin(x + Math.Sin(x + Math.Sin(x) / 3f) / 3f);
 		}
 
 		static void GenerateSampleFile(GeneratorFunc func, string path, int samples)
